Tests/LibWeb/Text/expected/HTML/document-write-after-close.txt

document.write after document.close reopens document: true
